+++ cryptol-saw-core/src/CryptolSAWCore/Cryptol.hs
@@ -4,6 +4,7 @@
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TemplateHaskell #-}
+{-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE ViewPatterns #-}
 
 {- |
@@ -74,6 +75,7 @@ import Prettyprinter ((<+>))
 -- cryptol
 import qualified Cryptol.Eval.Type as TV
 import qualified Cryptol.Backend as V
+import qualified Cryptol.Backend.FloatHelpers as V
 import qualified Cryptol.Backend.Monad as V
 import qualified Cryptol.Backend.SeqMap as V
 import qualified Cryptol.Backend.WordValue as V
@@ -103,6 +105,7 @@ import qualified Cryptol.ModuleSystem.Renamer as MR
 import qualified SAWSupport.Pretty as PPS
 
 -- saw-core
+import SAWCore.FiniteValue (FirstOrderFloat(..))
 import qualified SAWCore.Simulator.Concrete as SC
 import qualified SAWCore.Simulator.Value as SC
 import SAWCore.Prim (BitVector(..))
@@ -327,7 +330,7 @@ importType sc ty = do
             C.TCBit      -> scBoolType sc
             C.TCInteger  -> scIntegerType sc
             C.TCIntMod   -> scGlobalApply sc "Cryptol.IntModNum" =<< traverse go tyargs
-            C.TCFloat    -> scGlobalApply sc "Cryptol.TCFloat"   =<< traverse go tyargs
+            C.TCFloat    -> scGlobalApply sc "Cryptol.FloatNum"  =<< traverse go tyargs
             C.TCArray    -> do a <- go (tyargs !! 0)
                                b <- go (tyargs !! 1)
                                scArrayType sc a b
@@ -523,7 +526,7 @@ provePropRec sc prop0 prop = do
         (C.pIsZero -> Just (C.tIsFloat -> Just (e, p)))
           -> do e' <- importType sc e
                 p' <- importType sc p
-                scGlobalApply sc "Cryptol.PZeroFloat" [e', p']
+                scGlobalApply sc "Cryptol.PZeroFloatNum" [e', p']
         -- instance (Zero a) => Zero [n]a
         (C.pIsZero -> Just (C.tIsSeq -> Just (n, a)))
           -> do n' <- importType sc n
@@ -606,7 +609,7 @@ provePropRec sc prop0 prop = do
         (C.pIsRing -> Just (C.tIsFloat -> Just (e, p)))
           -> do e' <- importType sc e
                 p' <- importType sc p
-                scGlobalApply sc "Cryptol.PRingFloat" [e', p']
+                scGlobalApply sc "Cryptol.PRingFloatNum" [e', p']
         -- instance (Ring a) => Ring [n]a
         (C.pIsRing -> Just (C.tIsSeq -> Just (n, a)))
           -> do n' <- importType sc n
@@ -658,7 +661,7 @@ provePropRec sc prop0 prop = do
         (C.pIsField -> Just (C.tIsFloat -> Just (e, p)))
           -> do e' <- importType sc e
                 p' <- importType sc p
-                scGlobalApply sc "Cryptol.PFieldFloat" [e', p']
+                scGlobalApply sc "Cryptol.PFieldFloatNum" [e', p']
 
         -- instance Round Rational
         (C.pIsRound -> Just (C.tIsRational -> True))
@@ -667,7 +670,7 @@ provePropRec sc prop0 prop = do
         (C.pIsRound -> Just (C.tIsFloat -> Just (e, p)))
           -> do e' <- importType sc e
                 p' <- importType sc p
-                scGlobalApply sc "Cryptol.PRoundFloat" [e', p']
+                scGlobalApply sc "Cryptol.PRoundFloatNum" [e', p']
 
         -- instance Eq Bit
         (C.pIsEq -> Just (C.tIsBit -> True))
@@ -686,7 +689,7 @@ provePropRec sc prop0 prop = do
         (C.pIsEq -> Just (C.tIsFloat -> Just (e, p)))
           -> do e' <- importType sc e
                 p' <- importType sc p
-                scGlobalApply sc "Cryptol.PEqFloat" [e', p']
+                scGlobalApply sc "Cryptol.PEqFloatNum" [e', p']
         -- instance (fin n) => Eq [n]
         (C.pIsEq -> Just (C.tIsSeq -> Just (n, C.tIsBit -> True)))
           -> do n' <- importType sc n
@@ -736,7 +739,7 @@ provePropRec sc prop0 prop = do
         (C.pIsCmp -> Just (C.tIsFloat -> Just (e, p)))
           -> do e' <- importType sc e
                 p' <- importType sc p
-                scGlobalApply sc "Cryptol.PCmpFloat" [e', p']
+                scGlobalApply sc "Cryptol.PCmpFloatNum" [e', p']
         -- instance (fin n) => Cmp [n]
         (C.pIsCmp -> Just (C.tIsSeq -> Just (n, C.tIsBit -> True)))
           -> do n' <- importType sc n
@@ -834,7 +837,7 @@ provePropRec sc prop0 prop = do
         (C.pIsLiteral -> Just (_, C.tIsFloat -> Just (e, p)))
           -> do e' <- importType sc e
                 p' <- importType sc p
-                scGlobalApply sc "Cryptol.PLiteralFloat" [e', p']
+                scGlobalApply sc "Cryptol.PLiteralFloatNum" [e', p']
 
         -- instance (2 >= val) => LiteralLessThan val Bit
         (C.pIsLiteralLessThan -> Just (_, C.tIsBit -> True))
@@ -857,7 +860,7 @@ provePropRec sc prop0 prop = do
         (C.pIsLiteralLessThan -> Just (_, C.tIsFloat -> Just (e, p)))
           -> do e' <- importType sc e
                 p' <- importType sc p
-                scGlobalApply sc "Cryptol.PLiteralFloat" [e', p']
+                scGlobalApply sc "Cryptol.PLiteralFloatNum" [e', p']
 
         -- Note that in the FLiteral instances below, we intentionally do not
         -- translate the first three arguments.
@@ -870,7 +873,7 @@ provePropRec sc prop0 prop = do
         (C.pIsFLiteral -> Just (_, _, _, C.tIsFloat -> Just (e, p)))
           -> do e' <- importType sc e
                 p' <- importType sc p
-                scGlobalApply sc "Cryptol.PFLiteralFloat" [e', p']
+                scGlobalApply sc "Cryptol.PFLiteralFloatNum" [e', p']
 
         _ -> do
             let prop0' = "   " <> CryPP.pp prop0
@@ -1245,22 +1248,22 @@ floatPrims =
   , ("fpPosInf",       flip scGlobalDef "Cryptol.ecFpPosInf")
   , ("fpFromBits",     flip scGlobalDef "Cryptol.ecFpFromBits")
   , ("fpToBits",       flip scGlobalDef "Cryptol.ecFpToBits")
-  , ("=.=",            flip scGlobalDef "Cryptol.ecFpEq")
+  , ("=.=",            flip scGlobalDef "Cryptol.ecFpLogicalEq")
   , ("fpAdd",          flip scGlobalDef "Cryptol.ecFpAdd")
   , ("fpSub",          flip scGlobalDef "Cryptol.ecFpSub")
   , ("fpMul",          flip scGlobalDef "Cryptol.ecFpMul")
   , ("fpDiv",          flip scGlobalDef "Cryptol.ecFpDiv")
   , ("fpToRational",   flip scGlobalDef "Cryptol.ecFpToRational")
   , ("fpFromRational", flip scGlobalDef "Cryptol.ecFpFromRational")
-  , ("fpIsNaN",        flip scGlobalDef "Cryptol.fpIsNaN")
-  , ("fpIsInf",        flip scGlobalDef "Cryptol.fpIsInf")
-  , ("fpIsZero",       flip scGlobalDef "Cryptol.fpIsZero")
-  , ("fpIsNeg",        flip scGlobalDef "Cryptol.fpIsNeg")
-  , ("fpIsNormal",     flip scGlobalDef "Cryptol.fpIsNormal")
-  , ("fpIsSubnormal",  flip scGlobalDef "Cryptol.fpIsSubnormal")
-  , ("fpFMA",          flip scGlobalDef "Cryptol.fpFMA")
-  , ("fpAbs",          flip scGlobalDef "Cryptol.fpAbs")
-  , ("fpSqrt",         flip scGlobalDef "Cryptol.fpSqrt")
+  , ("fpIsNaN",        flip scGlobalDef "Cryptol.ecFpIsNaN")
+  , ("fpIsInf",        flip scGlobalDef "Cryptol.ecFpIsInf")
+  , ("fpIsZero",       flip scGlobalDef "Cryptol.ecFpIsZero")
+  , ("fpIsNeg",        flip scGlobalDef "Cryptol.ecFpIsNeg")
+  , ("fpIsNormal",     flip scGlobalDef "Cryptol.ecFpIsNormal")
+  , ("fpIsSubnormal",  flip scGlobalDef "Cryptol.ecFpIsSubnormal")
+  , ("fpFMA",          flip scGlobalDef "Cryptol.ecFpFMA")
+  , ("fpAbs",          flip scGlobalDef "Cryptol.ecFpAbs")
+  , ("fpSqrt",         flip scGlobalDef "Cryptol.ecFpSqrt")
   ]
 
 suiteBPrims :: Map C.PrimIdent (SharedContext -> IO Term)
@@ -2005,25 +2008,19 @@ proveEq sc t1 t2
   | otherwise =
     case (C.tNoUser t1, C.tNoUser t2) of
       (C.tIsSeq -> Just (n1, a1), C.tIsSeq -> Just (n2, a2)) ->
-        do n1' <- importType sc n1
-           n2' <- importType sc n2
+        do (n1', n2', nEq) <- proveNumEq sc n1 n2
            a1' <- importType sc a1
            a2' <- importType sc a2
-           num <- scGlobalApply sc "Cryptol.Num" []
-           nEq <- if n1 == n2
-                  then scGlobalApply sc "Prelude.Refl" [num, n1']
-                  else scGlobalApply sc "Prelude.unsafeAssert" [num, n1', n2']
            aEq <- proveEq sc a1 a2
            if a1 == a2
              then scGlobalApply sc "Cryptol.seq_cong1" [n1', n2', a1', nEq]
              else scGlobalApply sc "Cryptol.seq_cong" [n1', n2', a1', a2', nEq, aEq]
+      (C.tIsFloat -> Just (e1, p1), C.tIsFloat -> Just (e2, p2)) ->
+        do (e1', e2', eEq) <- proveNumEq sc e1 e2
+           (p1', p2', pEq) <- proveNumEq sc p1 p2
+           scGlobalApply sc "Cryptol.FloatNum_cong" [e1', e2', p1', p2', eEq, pEq]
       (C.tIsIntMod -> Just n1, C.tIsIntMod -> Just n2) ->
-        do n1' <- importType sc n1
-           n2' <- importType sc n2
-           num <- scGlobalApply sc "Cryptol.Num" []
-           nEq <- if n1 == n2
-                  then scGlobalApply sc "Prelude.Refl" [num, n1']
-                  else scGlobalApply sc "Prelude.unsafeAssert" [num, n1', n2']
+        do (n1', n2', nEq) <- proveNumEq sc n1 n2
            scGlobalApply sc "Cryptol.IntModNum_cong" [n1', n2', nEq]
       (C.tIsFun -> Just (a1, b1), C.tIsFun -> Just (a2, b2)) ->
         do a1' <- importType sc a1
@@ -2082,6 +2079,20 @@ proveEq sc t1 t2
             "t2: " <> CryPP.pp t2
         ]
 
+
+-- | Convert two @Num@ types to SAWCore 'Term's and create an equality term
+-- between them. If the two @Num@s are syntactically equal, this simply uses
+-- @Refl@ to equate them. Otherwise, this uses @unsafeAssert@, assuming that
+-- Cryptol's typechecker has proven that the two @Num@s are in fact equal.
+proveNumEq :: SharedContext -> C.Type -> C.Type -> IO (Term, Term, Term)
+proveNumEq sc n1 n2 =
+  do n1' <- importType sc n1
+     n2' <- importType sc n2
+     num <- scGlobalApply sc "Cryptol.Num" []
+     nEq <- if n1 == n2
+            then scGlobalApply sc "Prelude.Refl" [num, n1']
+            else scGlobalApply sc "Prelude.unsafeAssert" [num, n1', n2']
+     pure (n1', n2', nEq)
 
 -- | Create an equality term between two @RecordType@ values with the same
 -- field name. This works for equalities between record values and newtype
@@ -2390,6 +2401,7 @@ scCryptolType sc t =
       SC.VIntType -> return (Right C.tInteger)
       SC.VIntModType n -> return (Right (C.tIntMod (C.tNum n)))
       SC.VRationalType -> return (Right C.tRational)
+      SC.VFloatType e p -> pure (Right (C.tFloat (C.tNum e) (C.tNum p)))
       SC.VArrayType v1 v2 -> do
         Right t1 <- asCryptolTypeValue v1
         Right t2 <- asCryptolTypeValue v2
@@ -2467,7 +2479,15 @@ exportValue ty v = case ty of
       SC.VRational numer denom -> pure $ V.VRational $ V.SRational numer denom
       _ -> error $ "exportValue (on rational type " ++ show ty ++ ")"
 
-  TV.TVFloat _ _ -> panic "exportValue" ["Not yet implemented: Float"]
+  TV.TVFloat e p ->
+    case v of
+      SC.VFloat fof ->
+        pure $ V.VFloat $ V.BF
+          { V.bfExpWidth = e
+          , V.bfPrecWidth = p
+          , V.bfValue = fofValue fof
+          }
+      _ -> error $ "exportValue (on float type " ++ show ty ++ ")"
 
   TV.TVSeq _ e ->
     case v of